--- Runtime/Common/EditorUtilities.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9b4bf21bb108d1240beac30544e5525f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
